Toujours pourrie, essais avec Buffered Serial au lieu de Serial
  (on behalf of Luczia)

  modified:   Prg_Central_full/MotorCtrl/MotorCtrl.cpp
  modified:   Prg_Central_full/MotorCtrl/MotorCtrl.h
  modified:   Prg_Central_full/USBSerialCom/USBSerialCom.cpp
  modified:   Prg_Central_full/USBSerialCom/USBSerialCom.h
  modified:   Prg_Central_full/main.cpp

+++ Prg_Central_full/MotorCtrl/MotorCtrl.h
@@ -62,7 +62,7 @@ public:
     void SystemCtrl();
     void ResetCtrl();
     
-    void Debug(int orien, int dist);
+    void Debug(float orien, float dist);
 
     double getODO_X();
     double getODO_Y();
